+++ changes.md
@@ -1,6 +1,15 @@
 **Api changes**
 
 <details>
+<summary>Added Type(s)</summary>
+
+- added type `DiscountCombinationMode`
+- added type `DiscountsConfiguration`
+- added type `ProjectSetDiscountsConfigurationAction`
+</details>
+
+
+<details>
 <summary>Removed Property(s)</summary>
 
 - :warning: removed property `skipConfiguration` from type `RecurringOrderSetOrderSkipConfigurationAction`
@@ -10,6 +19,7 @@
 <details>
 <summary>Added Property(s)</summary>
 
+- added property `discounts` to type `Project`
 - added property `skipConfigurationInputDraft` to type `RecurringOrderSetOrderSkipConfigurationAction`
 </details>